+++ src/constants.js
@@ -1,0 +1,18 @@
+/**
+ * Constants object for the application.
+ * @module constants
+ */
+
+/**
+ * Constants object for application
+ */
+export default {
+  // Application name for angular
+  appName: 'angularjs-es6',
+
+  // API base url
+  restUrl: process.env.ANGULARJS_ES6_API_URL,
+
+  // Save the environment
+  env: process.env.NODE_ENV
+}